Added m2m relatioship between tasks and labels

+++ src/main/java/hexlet/code/app/sevice/TaskStatusService.java
@@ -5,11 +5,11 @@ import hexlet.code.app.dto.taskstatus.TaskStatusDTO;
 import hexlet.code.app.dto.taskstatus.TaskStatusUpdateDTO;
 import hexlet.code.app.exception.ResourceNotFoundException;
 import hexlet.code.app.mapper.taskstatus.TaskStatusMapper;
+import hexlet.code.app.repository.TaskRepository;
 import hexlet.code.app.repository.TaskStatusRepository;
+import org.apache.coyote.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -20,6 +20,9 @@ public class TaskStatusService {
 
     @Autowired
     private TaskStatusRepository taskStatusRepository;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     public List<TaskStatusDTO> findAll() {
         return taskStatusRepository.findAll()
@@ -49,13 +52,10 @@ public class TaskStatusService {
         return taskStatusMapper.map(model);
     }
 
-    public void destroy(Long id) {
-//        var model = taskStatusRepository.findById(id)
-//                .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
-//
-//        if (!model.getTasks().isEmpty()) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot delete task status with tasks");
-//        }
+    public void destroy(Long id) throws BadRequestException {
+        if (taskRepository.existsByTaskStatusId(id)) {
+            throw new BadRequestException("You can not delete status with task");
+        }
 
         taskStatusRepository.deleteById(id);
     }
